pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Mikhail <mixasibb@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.95.2"
uvicorn = "^0.22.0"
sqlalchemy = "^2.0.15"
pytest = "^7.3.1"
requests = "^2.31.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.6"
bcrypt = "^4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
